fix: правка особеенностей юнита
